h_testing

+++ vla-scripts/finetune_withHNN.py
@@ -1051,8 +1051,8 @@ def finetune(cfg: FinetuneConfig) -> None:
         nn.Linear(ACTION_DIM * 2, 64, bias=True),
         nn.ReLU(),
         nn.Linear(64,2,bias = True)
-    ).to(device_id)
-    time_d = Time_Derivative(2 * ACTION_DIM).to(device_id)
+    ).to(device_id).to(torch.bfloat16)
+    time_d = Time_Derivative(2 * ACTION_DIM).to(device_id).to(torch.bfloat16)
 
 
     # Instantiate optimizer
